Update type and code system objects to serialize without json mappings
