+++ src/main/java/org/opensrp/domain/CodeSystem.java
@@ -4,6 +4,7 @@
 package org.opensrp.domain;
 
 import java.io.Serializable;
+import java.util.List;
 
 /**
  * @author Samuel Githengi created on 08/30/19
@@ -12,34 +13,14 @@ public class CodeSystem implements Serializable {
 
 	private static final long serialVersionUID = -2587903025581183298L;
 
-	private String system;
+	private List<Code> coding;
 
-	private String code;
-
-	private String display;
-
-	public String getSystem() {
-		return system;
+	public List<Code> getCoding() {
+		return coding;
 	}
 
-	public void setSystem(String system) {
-		this.system = system;
-	}
-
-	public String getCode() {
-		return code;
-	}
-
-	public void setCode(String code) {
-		this.code = code;
-	}
-
-	public String getDisplay() {
-		return display;
-	}
-
-	public void setDisplay(String display) {
-		this.display = display;
+	public void setCoding(List<Code> coding) {
+		this.coding = coding;
 	}
 
 }
